Dump build info into build directory
Since the build directory is a hash of requirement and user config,
it is very hard to inspect which build directory is for what library.
In order to improve inspection, in each build directory,
a build info json file is dumped, with requirement information,
as well as user provided configuration (via cli args).
This will help with inspecting the build directories, as well as
creating future tooling that operates on them.