Add initial git url support
Previously, download worked only for .tar.gz urls.
Add support for user providing the git url of the repository.
Since, now the download has to handle two diffrent
urls, check how the url ends, and throw exception,
if it does not satisfy the requirements.
Additionally there is a requirement, that the git repository
must contain a tag/branch named after the version.
Removed one of the TODO's matching this commit.
Increment the alfa version of the project.
Update readme accordingly.