Replace calculated install directories
Modify where how build and install directories are named.
The reason for change was scope definition.
This is not a package manager, it is simply a tool
to speed up prototyping of your projects.
Now, there is only one scheme of build directory naming,
and it is {name}-{version}, just so it can be descriptive.
The install location is the same, no matter the user configuration.
To accomodate for multiple build types (Release, Debug) an
argument was introduced, that allows the user to configure build
directory.