Add initial version of the schema
Enforce json file corectness by introducing json schema to the project.
Upon launch, the loaded json file is validated with newly introduced
schema. Upon any error, the program aborts.
This will help with removing a lot of not needed checks in the future.
Another benefit is that the difference between syntax of each
version is now much more visible.