Add quicktype generated code from schema
Quicktype is a service that allows to generate models
from json schema.
As this is a web service, and does not allow to automatically
generate the classes, commit the generated code as quicktype.py
file.

Those classes will be used in future changes as a base for working
with requirements file.
This will allow for proper cleanup of functions, and remove any
discrepencies between parsing, and schema.