Switch requirement.install to free function
The install method of requirement made it hard to make swap
to the quicktype generated classes.
To prepare for that move the function as a free func, and rename it to
install_requirement.
Replace any self reference to the "requirement" argument which was added
instead of self.

Replace any usage of the method, with the free function.