Reparametrize download function
Change the function arguments, to accept only requirement_handler.
In order to not change the function much, create within the function two
variables that were previously the arguments.
Assign to the variables values that are kept inside requirement_handler.

Fix the only invokation of the download function in the codebase