Add argparse dependency
The project will need to parse arguments provided during invokation.
To do that the argparse hader only library will be used.
Add a cmake commands to download the dependency and modify the main.cpp
file to use the library in most simple manner.
The usage of the library is not meant to provide any value other
than to show that the program compiles and links with this library.