Add nlohmann json library
The iwd parsers json file that describes all dependencies
that will be used to build the active target.
In order to provided the json parsing nlohmann json library
shall be used.
Add the project to the cmake, and link the nlohmann json library
with produced target.
The modification in the main cpp file are only mean to verify
that the iwd application can compile and link while using nlohmann
headers.