Add LibArchive requirement
Rely on externally installed LibArchive. Use find_package
in order to declare dependency. The LibArchive package can be
found in most of the linux systems so it is ok to include it this way.
LibArchive will be needed to handle decompressing of
targz files that can be specified as a source of the requirement.