Add arguments parsing
Firstly, create the iwd directory and namespace.
Set project include directories, so there is no relative include
happening.
Then create arguments structure that can be easly read from.

Create parse_args function that returns the arguments structure and
accepts argc, argv as arguments. The function parses the arguments
in an almost compatible way with the python version.
The argument parser for each argument has an action specified that
sets the field of arguments structure. This is done so
there can be clean separation from the library. In the future
this library may be changed for something more custom or roboust.