Add directory structure
Add directory structure into the project.
The directory is a small wrapper around std::fs::path that
exposes easy to use methods and forces the user to explicitly
convert the path object into directory.
It helps to enforce clean api.

The ctor of directory has assertion that verifies whether
the path specified points to a directory.

Additionally the iterator_range structure was introduced that is a
holder of two iterators. This is mean to allow to return
iterator pairs easier.

Everything is placed within namespace vn, as:
1) This should be isolated from actual iwd code
2) This will be moved to external repository (vn utility library)