Create check_call function
This function is responsible for launching the subprocess
with given arguments and settings and waiting for it.
Much like the python eqivalent, when the subprocess returns
with error (non 0 exit code), the function throws runtime_error.

By default the function pipes its outputs to main process ones.

As the arguments for this function will grow, they are grouped
in check_call_arguments structure.

For now there are two arguments:
Timeout argument, when set makes the check_call function
throw an error if the process did not exit after specified time.
Working directory argument, when specified changes the
working directory from current.

Added Boost system and filesystem dependency

*NOTE* The filesystem dependency is because of boost::process, it
internally uses the boost::filesystem.
The api is wrapped, so the boost::filesystem usage should not go outside
of check_call functionality.

Added Threads dependency