Add requirement handler class
Add the class that holds both the quicktype requirement
and directories.
This class will be responsible for executing requirements tasks
like downloading/configuring and installing.

For now add initial implementation of download function that
only downloads and does not extract sources.
It also verifies whether the sources are already downloaded.

In main add example usage.