Add extract_tarfile function
This function uses libarchive to extract the
archive into specified directory, additionally
in accepts the functor that returns boolean value.
True value returned means that the entry should be extracted.
Additionally as an input for the functor is a reference to tar file.
It can be used to modify the filename of the extracted file.