Add glob_recursive function
This initial implementation of glob function
accepts two arguments,
target directory that should it recrusively scan files
and the pattern, that is an fnmatch pattern, that will specify which
files should be shown.

As per glob requirements, the function does not list hidden files
(even if the pattern starts with dot), and it does not recurse
into hidden directories.