Add TODO into requirement_handler download function
Since there is no issue managment, all important things that
need to be improved are tracked in TODOs that exist all around the
project.

This one touches the improvement, where when we download targz
file, the actual contents of it are sometimes wrapped in additional
directory.
This case could be detected during extraction, and the contents of it
could be extracted directly into iwd source dir.