Create executable class
Instead of calling processes directly, with assumption that
they are installed in the users machine, create executable class
that represents existing path to target process that can be run.

Provide try_find_executable function that returns optional
of this executable.

Main idea is to call try_find_ex functions at the very begginging
in order to ensure that all executables such as git and cmake can be
found and launched.