Create cmake struct
The struct is a wrapper around check_calls that provides
configure and install functions.
This will be used as replacement to plain executable calls in
requirement.

The create function provides initial checks, whether the executable
that is actually cmake executable.

Additionally, as the version() is exposed, this will serve as a checker
for minimal cmake version that the iwd supports.